@@ -1,66 +1,75 @@
-﻿using MessagePack;
-using MessagePack.Formatters;
-using MessagePack.Internal;
-using MessagePack.Resolvers;
-using SharedData;
+﻿// Copyright (c) All contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
 using System;
+using System.Buffers;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.IO;
 using System.Linq;
 using System.Reflection.Emit;
 using System.Text;
 using System.Threading.Tasks;
+using MessagePack;
+using MessagePack.Formatters;
+using MessagePack.Internal;
+using MessagePack.Resolvers;
+using Nerdbank.Streams;
+using SharedData;
 
 namespace DynamicCodeDumper
 {
     public class Program
     {
-        static void Main(string[] args)
+        private static void Main(string[] args)
         {
             try
             {
                 // var mi = AutomataKeyGen.GetGetKeyMethod();
-                // mi.Invoke(null, new[] { 
+                // mi.Invoke(null, new[] {
 
-                //DynamicObjectResolver.Instance.GetFormatter<ArrayOptimizeClass>();
-                //DynamicObjectResolver.Instance.GetFormatter<Empty1>();
-                //DynamicObjectResolver.Instance.GetFormatter<Empty2>();
+                ////DynamicObjectResolver.Instance.GetFormatter<ArrayOptimizeClass>();
+                ////DynamicObjectResolver.Instance.GetFormatter<Empty1>();
+                ////DynamicObjectResolver.Instance.GetFormatter<Empty2>();
 
-                //DynamicObjectResolver.Instance.GetFormatter<NonEmpty1>();
-                //DynamicObjectResolver.Instance.GetFormatter<NonEmpty2>();
-                //DynamicObjectResolver.Instance.GetFormatter<FirstSimpleData>();
-                //DynamicObjectResolver.Instance.GetFormatter<Version0>();
-                //DynamicObjectResolver.Instance.GetFormatter<Version1>();
-                //DynamicObjectResolver.Instance.GetFormatter<Version2>();
-                //DynamicObjectResolver.Instance.GetFormatter<SimpleIntKeyData>();
-                //DynamicObjectResolver.Instance.GetFormatter<SimlpeStringKeyData>();
-                //DynamicObjectResolver.Instance.GetFormatter<SimlpeStringKeyData2>();
-                //DynamicObjectResolver.Instance.GetFormatter<StringKeySerializerTarget>();
-                //DynamicObjectResolver.Instance.GetFormatter<LongestString>();
-                var f = DynamicObjectResolverAllowPrivate.Instance.GetFormatter<MyClass>();
-                //DynamicObjectResolver.Instance.GetFormatter<StringKeySerializerTargetBinary>();
-                //DynamicObjectResolver.Instance.GetFormatter<Callback1>();
-                //DynamicObjectResolver.Instance.GetFormatter<Callback1_2>();
-                //DynamicObjectResolver.Instance.GetFormatter<Callback2>();
-                //DynamicObjectResolver.Instance.GetFormatter<Callback2_2>();
+                ////DynamicObjectResolver.Instance.GetFormatter<NonEmpty1>();
+                ////DynamicObjectResolver.Instance.GetFormatter<NonEmpty2>();
+                ////DynamicObjectResolver.Instance.GetFormatter<FirstSimpleData>();
+                ////DynamicObjectResolver.Instance.GetFormatter<Version0>();
+                ////DynamicObjectResolver.Instance.GetFormatter<Version1>();
+                ////DynamicObjectResolver.Instance.GetFormatter<Version2>();
+                ////DynamicObjectResolver.Instance.GetFormatter<SimpleIntKeyData>();
+                ////DynamicObjectResolver.Instance.GetFormatter<SimpleStringKeyData>();
+                ////DynamicObjectResolver.Instance.GetFormatter<SimpleStringKeyData2>();
+                ////DynamicObjectResolver.Instance.GetFormatter<StringKeySerializerTarget>();
+                ////DynamicObjectResolver.Instance.GetFormatter<LongestString>();
+                IMessagePackFormatter<MyClass> f = DynamicObjectResolverAllowPrivate.Instance.GetFormatter<MyClass>();
+                ////DynamicObjectResolver.Instance.GetFormatter<StringKeySerializerTargetBinary>();
+                ////DynamicObjectResolver.Instance.GetFormatter<Callback1>();
+                ////DynamicObjectResolver.Instance.GetFormatter<Callback1_2>();
+                ////DynamicObjectResolver.Instance.GetFormatter<Callback2>();
+                ////DynamicObjectResolver.Instance.GetFormatter<Callback2_2>();
 
-                //DynamicUnionResolver.Instance.GetFormatter<IHogeMoge>();
-                //DynamicUnionResolver.Instance.GetFormatter<IUnionChecker>();
-                //DynamicUnionResolver.Instance.GetFormatter<IUnionChecker2>();
+                ////DynamicUnionResolver.Instance.GetFormatter<IHogeMoge>();
+                ////DynamicUnionResolver.Instance.GetFormatter<IUnionChecker>();
+                ////DynamicUnionResolver.Instance.GetFormatter<IUnionChecker2>();
 
-                //DynamicUnionResolver.Instance.GetFormatter<RootUnionType>();
+                ////DynamicUnionResolver.Instance.GetFormatter<RootUnionType>();
 
-                //DynamicEnumResolver.Instance.GetFormatter<IntEnum>();
-                //DynamicEnumResolver.Instance.GetFormatter<ShortEnum>();
+                ////DynamicEnumResolver.Instance.GetFormatter<IntEnum>();
+                ////DynamicEnumResolver.Instance.GetFormatter<ShortEnum>();
 
-                //DynamicContractlessObjectResolver.Instance.GetFormatter<ContractlessConstructorCheck>();
-                //DynamicContractlessObjectResolver.Instance.GetFormatter<Contractless2>();
+                ////DynamicContractlessObjectResolver.Instance.GetFormatter<ContractlessConstructorCheck>();
+                ////DynamicContractlessObjectResolver.Instance.GetFormatter<Contractless2>();
 
-                //DynamicContractlessObjectResolver.Instance.GetFormatter<EntityBase>();
+                ////DynamicContractlessObjectResolver.Instance.GetFormatter<EntityBase>();
 
-                byte[] b = null;
-                var bin = f.Serialize(ref b, 0, new MyClass { MyProperty1 = 100, MyProperty2 = "foo" }, null);
-
+                using (var sequence = new Sequence<byte>())
+                {
+                    var sequenceWriter = new MessagePackWriter(sequence);
+                    f.Serialize(ref sequenceWriter, new MyClass { MyProperty1 = 100, MyProperty2 = "foo" }, EmptyResolver.Options);
+                    sequenceWriter.Flush();
+                }
             }
             catch (Exception ex)
             {
@@ -68,22 +77,22 @@ namespace DynamicCodeDumper
             }
             finally
             {
-                var a1 = DynamicObjectResolver.Instance.Save();
-                var a2 = DynamicUnionResolver.Instance.Save();
-                var a3 = DynamicEnumResolver.Instance.Save();
-                var a4 = DynamicContractlessObjectResolver.Instance.Save();
-                //var a5 = AutomataKeyGen.Save();
+                AssemblyBuilder a1 = DynamicObjectResolver.Instance.Save();
+                AssemblyBuilder a2 = DynamicUnionResolver.Instance.Save();
+                AssemblyBuilder a3 = DynamicEnumResolver.Instance.Save();
+                AssemblyBuilder a4 = DynamicContractlessObjectResolver.Instance.Save();
+                ////var a5 = AutomataKeyGen.Save();
 
-                //Verify(a5);
+                ////Verify(a5);
             }
-            //Verify(a1, a2, a3, a4);
+            ////Verify(a1, a2, a3, a4);
         }
 
-        static void Verify(params AssemblyBuilder[] builders)
+        private static void Verify(params AssemblyBuilder[] builders)
         {
-            var path = @"C:\Program Files (x86)\Microsoft SDKs\Windows\v10.0A\bin\NETFX 4.6.1 Tools\x64\PEVerify.exe";
+            var path = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ProgramFilesX86), @"Microsoft SDKs\Windows\v10.0A\bin\NETFX 4.6.1 Tools\x64\PEVerify.exe");
 
-            foreach (var targetDll in builders)
+            foreach (AssemblyBuilder targetDll in builders)
             {
                 var psi = new ProcessStartInfo(path, targetDll.GetName().Name + ".dll")
                 {
@@ -91,7 +100,7 @@ namespace DynamicCodeDumper
                     WindowStyle = ProcessWindowStyle.Hidden,
                     RedirectStandardOutput = true,
                     RedirectStandardError = true,
-                    UseShellExecute = false
+                    UseShellExecute = false,
                 };
 
                 var p = Process.Start(psi);
@@ -99,54 +108,66 @@ namespace DynamicCodeDumper
                 Console.WriteLine(data);
             }
         }
+
+        private class EmptyResolver : IFormatterResolver
+        {
+            internal static readonly MessagePackSerializerOptions Options = new MessagePackSerializerOptions(new EmptyResolver());
+
+            public IMessagePackFormatter<T> GetFormatter<T>() => null;
+        }
     }
+
     [MessagePackObject]
     public class MyClass
     {
         [Key(0)]
         [MessagePackFormatter(typeof(Int_x10Formatter))]
         public int MyProperty1 { get; set; }
+
         [Key(1)]
         [MessagePackFormatter(typeof(String_x2Formatter))]
         public string MyProperty2 { get; set; }
 
+#pragma warning disable SA1306 // Field names should begin with lower-case letter
         [Key(2)]
         private int Foo;
+#pragma warning restore SA1306 // Field names should begin with lower-case letter
 
         public void SetFoo(int f)
         {
-            Foo = f;
+            this.Foo = f;
         }
 
         public int GetFoo()
         {
-            return Foo;
+            return this.Foo;
         }
     }
+
     public class Int_x10Formatter : IMessagePackFormatter<int>
     {
-        public int Deserialize(byte[] bytes, int offset, IFormatterResolver formatterResolver, out int readSize)
+        public int Deserialize(ref MessagePackReader reader, MessagePackSerializerOptions options)
         {
-            return MessagePackBinary.ReadInt32(bytes, offset, out readSize) * 10;
+            return reader.ReadInt32() * 10;
         }
 
-        public int Serialize(ref byte[] bytes, int offset, int value, IFormatterResolver formatterResolver)
+        public void Serialize(ref MessagePackWriter writer, int value, MessagePackSerializerOptions options)
         {
-            return MessagePackBinary.WriteInt32(ref bytes, offset, value * 10);
+            writer.WriteInt32(value * 10);
         }
     }
 
     public class String_x2Formatter : IMessagePackFormatter<string>
     {
-        public string Deserialize(byte[] bytes, int offset, IFormatterResolver formatterResolver, out int readSize)
+        public string Deserialize(ref MessagePackReader reader, MessagePackSerializerOptions options)
         {
-            var s = MessagePackBinary.ReadString(bytes, offset, out readSize);
+            var s = reader.ReadString();
             return s + s;
         }
 
-        public int Serialize(ref byte[] bytes, int offset, string value, IFormatterResolver formatterResolver)
+        public void Serialize(ref MessagePackWriter writer, string value, MessagePackSerializerOptions options)
         {
-            return MessagePackBinary.WriteString(ref bytes, offset, value + value);
+            writer.Write(value + value);
         }
     }
 
@@ -158,15 +179,14 @@ namespace DynamicCodeDumper
 
         public static bool CalledAfter = false;
 
-
         public Callback2_2(int x)
             : this(x, () => { }, () => { })
         {
-
         }
 
-        Action onBefore;
-        Action onAfter;
+        private Action onBefore;
+        private Action onAfter;
+
         public Callback2_2(int x, Action onBefore, Action onAfter)
         {
             this.X = x;
@@ -176,7 +196,7 @@ namespace DynamicCodeDumper
 
         void IMessagePackSerializationCallbackReceiver.OnBeforeSerialize()
         {
-            onBefore();
+            this.onBefore();
         }
 
         void IMessagePackSerializationCallbackReceiver.OnAfterDeserialize()
@@ -199,16 +219,15 @@ namespace DynamicCodeDumper
     {
     }
 
-
     public class EmptyContractless
     {
-
     }
 
     [MessagePackObject(true)]
-    public class SimlpeStringKeyData2
+    public class SimpleStringKeyData2
     {
         public int MyProperty1 { get; set; }
+
         public int MyProperty2 { get; set; }
     }
 
@@ -216,13 +235,21 @@ namespace DynamicCodeDumper
     public class StringKeySerializerTarget
     {
         public int MyProperty1 { get; set; }
+
         public int MyProperty2 { get; set; }
+
         public int MyProperty3 { get; set; }
+
         public int MyProperty4 { get; set; }
+
         public int MyProperty5 { get; set; }
+
         public int MyProperty6 { get; set; }
+
         public int MyProperty7 { get; set; }
+
         public int MyProperty8 { get; set; }
+
         public int MyProperty9 { get; set; }
     }
 
@@ -230,30 +257,53 @@ namespace DynamicCodeDumper
     public class StringKeySerializerTargetBinary
     {
         public int MyProperty12 { get; set; }
+
         public int MyProperty6 { get; set; }
+
         public int MyProperty45 { get; set; }
+
         public int MyProperty14 { get; set; }
+
         public int MyProperty7 { get; set; }
+
         public int MyProperty72 { get; set; }
+
         public int MyProperty99 { get; set; }
+
         public int MyProperty88 { get; set; }
 
         public int MyProperty1 { get; set; }
+
         public int MyProperty2 { get; set; }
+
         public int MyProperty5 { get; set; }
+
         public int MyProperty8 { get; set; }
+
         public int MyProperty9 { get; set; }
+
         public int MyProperty70 { get; set; }
+
         public int MyProperty16 { get; set; }
+
         public int MyProperty16MyProperty16MyProperty16MyProperty16 { get; set; }
+
         public int MyProperty23 { get; set; }
+
         public int MyProperty3 { get; set; }
+
         public int MyProperty10 { get; set; }
+
         public int MyProperty67 { get; set; }
+
         public int MyProperty68 { get; set; }
+
         public int MyProperty4 { get; set; }
+
         public int FooPropertyWithBazBaz { get; set; }
+
         public int MyProperty69 { get; set; }
+
         public int MyProperty71 { get; set; }
     }
 
@@ -261,8 +311,11 @@ namespace DynamicCodeDumper
     public class LongestString
     {
         public int MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1 { get; set; }
+
         public int MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty2 { get; set; }
+
         public int MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty1MyProperty2MyProperty { get; set; }
+
         public int OAFADFZEWFSDFSDFKSLJFWEFNWOZFUSEWWEFWEWFFFFFFFFFFFFFFZFEWBFOWUEGWHOUDGSOGUDSZNOFRWEUFWGOWHOGHWOG000000000000000000000000000000000000000HOGZ { get; set; }
     }
 
@@ -270,16 +323,18 @@ namespace DynamicCodeDumper
     public class Dup
     {
         public int ABCDEFGH { get; set; }
+
         public int ABCDEFGHIJKL { get; set; }
+
         public int ABCDEFGHIJKO { get; set; }
     }
 
     public class Contractless2
     {
         public int MyProperty { get; set; }
+
         public string MyProperty2 { get; set; }
     }
-
 
     public interface IEntity
     {
@@ -290,7 +345,7 @@ namespace DynamicCodeDumper
     {
         public Event(string name)
         {
-            Name = name;
+            this.Name = name;
         }
 
         public string Name { get; }
@@ -300,7 +355,7 @@ namespace DynamicCodeDumper
     {
         public Holder(IEntity entity)
         {
-            Entity = entity;
+            this.Entity = entity;
         }
 
         public IEntity Entity { get; }
@@ -312,102 +367,100 @@ namespace DynamicCodeDumper
 
         public EntityBase()
         {
-
         }
     }
 
-    //public class SimulateDup : MessagePack.Formatters.IMessagePackFormatter<Dup>
-    //{
-    //    public unsafe Dup Deserialize(byte[] bytes, int offset, IFormatterResolver formatterResolver, out int readSize)
-    //    {
-    //        if (MessagePackBinary.IsNil(bytes, offset))
-    //        {
-    //            readSize = 1;
-    //            return null;
-    //        }
+    ////public class SimulateDup : MessagePack.Formatters.IMessagePackFormatter<Dup>
+    ////{
+    ////    public unsafe Dup Deserialize(byte[] bytes, int offset, IFormatterResolver formatterResolver, out int readSize)
+    ////    {
+    ////        if (MessagePackBinary.IsNil(bytes, offset))
+    ////        {
+    ////            readSize = 1;
+    ////            return null;
+    ////        }
 
-    //        var startOffset = offset;
+    ////        var startOffset = offset;
 
-    //        var len = MessagePackBinary.ReadMapHeader(bytes, offset, out readSize);
-    //        offset += readSize;
+    ////        var len = MessagePackBinary.ReadMapHeader(bytes, offset, out readSize);
+    ////        offset += readSize;
 
+    ////        int aBCDEFGH = 0;
+    ////        int aBCDEFGHIJKL = 0;
+    ////        int aBCDEFGHIJKO = 0;
 
-    //        int aBCDEFGH = 0;
-    //        int aBCDEFGHIJKL = 0;
-    //        int aBCDEFGHIJKO = 0;
+    ////        // ---isStringKey
 
-    //        // ---isStringKey
+    ////        ulong key;
+    ////        ArraySegment<byte> arraySegment;
+    ////        byte* p;
+    ////        int rest;
 
-    //        ulong key;
-    //        ArraySegment<byte> arraySegment;
-    //        byte* p;
-    //        int rest;
+    ////        fixed (byte* buffer = &bytes[0])
+    ////        {
+    ////            for (int i = 0; i < len; i++)
+    ////            {
+    ////                arraySegment = MessagePackBinary.ReadStringSegment(bytes, offset, out readSize);
+    ////                offset += readSize;
 
-    //        fixed (byte* buffer = &bytes[0])
-    //        {
-    //            for (int i = 0; i < len; i++)
-    //            {
-    //                arraySegment = MessagePackBinary.ReadStringSegment(bytes, offset, out readSize);
-    //                offset += readSize;
+    ////                p = buffer + arraySegment.Offset;
+    ////                rest = arraySegment.Count;
 
-    //                p = buffer + arraySegment.Offset;
-    //                rest = arraySegment.Count;
+    ////                if (rest == 0) goto LOOP_END;
 
-    //                if (rest == 0) goto LOOP_END;
+    ////                key = AutomataKeyGen.GetKey(ref p, ref rest);
+    ////                if (rest == 0)
+    ////                {
+    ////                    if (key == 5208208757389214273L)
+    ////                    {
+    ////                        aBCDEFGH = MessagePackBinary.ReadInt32(bytes, offset, out readSize);
+    ////                        goto LOOP_END;
+    ////                    }
 
-    //                key = AutomataKeyGen.GetKey(ref p, ref rest);
-    //                if (rest == 0)
-    //                {
-    //                    if (key == 5208208757389214273L)
-    //                    {
-    //                        aBCDEFGH = MessagePackBinary.ReadInt32(bytes, offset, out readSize);
-    //                        goto LOOP_END;
-    //                    }
+    ////                    goto READ_NEXT;
+    ////                }
 
-    //                    goto READ_NEXT;
-    //                }
+    ////                if (key == 5208208757389214273L)
+    ////                {
+    ////                    key = AutomataKeyGen.GetKey(ref p, ref rest);
+    ////                    if (rest == 0)
+    ////                    {
+    ////                        if (key == 1280002633L)
+    ////                        {
+    ////                            aBCDEFGHIJKL = MessagePackBinary.ReadInt32(bytes, offset, out readSize);
+    ////                            goto LOOP_END;
+    ////                        }
 
-    //                if (key == 5208208757389214273L)
-    //                {
-    //                    key = AutomataKeyGen.GetKey(ref p, ref rest);
-    //                    if (rest == 0)
-    //                    {
-    //                        if (key == 1280002633L)
-    //                        {
-    //                            aBCDEFGHIJKL = MessagePackBinary.ReadInt32(bytes, offset, out readSize);
-    //                            goto LOOP_END;
-    //                        }
+    ////                        if (key == 1330334281L)
+    ////                        {
+    ////                            aBCDEFGHIJKO = MessagePackBinary.ReadInt32(bytes, offset, out readSize);
+    ////                            goto LOOP_END;
+    ////                        }
+    ////                    }
+    ////                }
 
-    //                        if (key == 1330334281L)
-    //                        {
-    //                            aBCDEFGHIJKO = MessagePackBinary.ReadInt32(bytes, offset, out readSize);
-    //                            goto LOOP_END;
-    //                        }
-    //                    }
-    //                }
+    ////                READ_NEXT:
+    ////                readSize = MessagePackBinary.ReadNextBlock(bytes, offset);
 
-    //                READ_NEXT:
-    //                readSize = MessagePackBinary.ReadNextBlock(bytes, offset);
+    ////                LOOP_END:
+    ////                offset += readSize;
+    ////                continue;
+    ////            }
+    ////        }
 
-    //                LOOP_END:
-    //                offset += readSize;
-    //                continue;
-    //            }
-    //        }
+    ////        // --- end
 
-    //        // --- end
+    ////        return new Dup
+    ////        {
+    ////            ABCDEFGH = aBCDEFGH,
+    ////            ABCDEFGHIJKL = aBCDEFGHIJKL,
+    ////            ABCDEFGHIJKO = aBCDEFGHIJKO
+    ////        };
+    ////    }
 
-    //        return new Dup
-    //        {
-    //            ABCDEFGH = aBCDEFGH,
-    //            ABCDEFGHIJKL = aBCDEFGHIJKL,
-    //            ABCDEFGHIJKO = aBCDEFGHIJKO
-    //        };
-    //    }
-
-    //    public int Serialize(ref byte[] bytes, int offset, Dup value, IFormatterResolver formatterResolver)
-    //    {
-    //        throw new NotImplementedException();
-    //    }
-    //}
+    ////    public int Serialize(ref byte[] bytes, int offset, Dup value, IFormatterResolver formatterResolver)
+    ////    {
+    ////        throw new NotImplementedException();
+    ////    }
+    ////}
 }
