@@ -1,56 +1,54 @@
-﻿using MessagePack.Formatters;
-using System.Collections.Immutable;
+﻿// Copyright (c) All contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
 using System;
+using System.Buffers;
+using System.Collections.Immutable;
+using MessagePack.Formatters;
+
+#pragma warning disable SA1649 // File name should match first type name
 
 namespace MessagePack.ImmutableCollection
 {
     // Immutablearray<T>.Enumerator is 'not' IEnumerator<T>, can't use abstraction layer.
     public class ImmutableArrayFormatter<T> : IMessagePackFormatter<ImmutableArray<T>>
     {
-        public int Serialize(ref byte[] bytes, int offset, ImmutableArray<T> value, IFormatterResolver formatterResolver)
+        public void Serialize(ref MessagePackWriter writer, ImmutableArray<T> value, MessagePackSerializerOptions options)
         {
             if (value == null)
             {
-                return MessagePackBinary.WriteNil(ref bytes, offset);
+                writer.WriteNil();
             }
             else
             {
-                var startOffset = offset;
-                var formatter = formatterResolver.GetFormatterWithVerify<T>();
+                IMessagePackFormatter<T> formatter = options.Resolver.GetFormatterWithVerify<T>();
 
-                offset += MessagePackBinary.WriteArrayHeader(ref bytes, offset, value.Length);
+                writer.WriteArrayHeader(value.Length);
 
-                foreach (var item in value)
+                foreach (T item in value)
                 {
-                    offset += formatter.Serialize(ref bytes, offset, item, formatterResolver);
+                    formatter.Serialize(ref writer, item, options);
                 }
-
-                return offset - startOffset;
             }
         }
 
-        public ImmutableArray<T> Deserialize(byte[] bytes, int offset, IFormatterResolver formatterResolver, out int readSize)
+        public ImmutableArray<T> Deserialize(ref MessagePackReader reader, MessagePackSerializerOptions options)
         {
-            if (MessagePackBinary.IsNil(bytes, offset))
+            if (reader.TryReadNil())
             {
-                readSize = 1;
                 return ImmutableArray<T>.Empty;
             }
             else
             {
-                var startOffset = offset;
-                var formatter = formatterResolver.GetFormatterWithVerify<T>();
+                IMessagePackFormatter<T> formatter = options.Resolver.GetFormatterWithVerify<T>();
 
-                var len = MessagePackBinary.ReadArrayHeader(bytes, offset, out readSize);
-                offset += readSize;
+                var len = reader.ReadArrayHeader();
 
-                var builder = ImmutableArray.CreateBuilder<T>(len);
+                ImmutableArray<T>.Builder builder = ImmutableArray.CreateBuilder<T>(len);
                 for (int i = 0; i < len; i++)
                 {
-                    builder.Add(formatter.Deserialize(bytes, offset, formatterResolver, out readSize));
-                    offset += readSize;
+                    builder.Add(formatter.Deserialize(ref reader, options));
                 }
-                readSize = offset - startOffset;
 
                 return builder.ToImmutable();
             }
@@ -182,7 +180,7 @@ namespace MessagePack.ImmutableCollection
 
         protected override ImmutableQueue<T> Complete(ImmutableQueueBuilder<T> intermediateCollection)
         {
-            return intermediateCollection.q;
+            return intermediateCollection.Q;
         }
 
         protected override ImmutableQueueBuilder<T> Create(int count)
@@ -273,7 +271,7 @@ namespace MessagePack.ImmutableCollection
 
         protected override IImmutableQueue<T> Complete(ImmutableQueueBuilder<T> intermediateCollection)
         {
-            return intermediateCollection.q;
+            return intermediateCollection.Q;
         }
 
         protected override ImmutableQueueBuilder<T> Create(int count)
@@ -300,16 +298,14 @@ namespace MessagePack.ImmutableCollection
         }
     }
 
-
     // pseudo builders
-
     public class ImmutableQueueBuilder<T>
     {
-        public ImmutableQueue<T> q = ImmutableQueue<T>.Empty;
+        public ImmutableQueue<T> Q { get; set; } = ImmutableQueue<T>.Empty;
 
         public void Add(T value)
         {
-            q = q.Enqueue(value);
+            this.Q = this.Q.Enqueue(value);
         }
     }
 }
