@@ -1,21 +1,24 @@
-﻿using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CodeFixes;
-using Microsoft.CodeAnalysis.Diagnostics;
+﻿// Copyright (c) All contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CodeFixes;
+using Microsoft.CodeAnalysis.Diagnostics;
 
 namespace MessagePackAnalyzer
 {
     // Store multiple errors.
     internal class DiagnosticsReportContext
     {
-        readonly List<Diagnostic> diagnostics = new List<Diagnostic>();
-        readonly SyntaxNodeAnalysisContext context;
+        private readonly List<Diagnostic> diagnostics = new List<Diagnostic>();
+        private readonly SyntaxNodeAnalysisContext context;
 
-        public IReadOnlyList<Diagnostic> Diagnostics => diagnostics;
+        public IReadOnlyList<Diagnostic> Diagnostics => this.diagnostics;
 
         public IReadOnlyList<string> AdditionalAllowTypes { get; }
 
@@ -27,14 +30,14 @@ namespace MessagePackAnalyzer
 
         public void Add(Diagnostic diagnostic)
         {
-            diagnostics.Add(diagnostic);
+            this.diagnostics.Add(diagnostic);
         }
 
         public void ReportAll()
         {
-            foreach (var item in diagnostics)
+            foreach (Diagnostic item in this.diagnostics)
             {
-                context.ReportDiagnostic(item);
+                this.context.ReportDiagnostic(item);
             }
         }
     }
